Rename Version_control.py to app_temp.py - Add superuser - Refactor some code who use user
