+++ Right Posture/modules/ui_main_functions.py
@@ -17,7 +17,7 @@
 # MAIN FILE
 # ///////////////////////////////////////////////////////////////
 from main import *
-from modules.Version_control import version
+from modules.app_temp import version
 # GLOBALS
 # ///////////////////////////////////////////////////////////////
 GLOBAL_STATE = False
