+++ Right Posture/setup.py
@@ -1,7 +1,7 @@
 import sys
 import os
 from cx_Freeze import setup, Executable
-from modules.Version_control import version
+from modules.app_temp import version
 
 Compile_version = version.thisVersion
 if Compile_version is None:
